Ref: #5643.
Summary: Give the ability to match the display refresh rate with the video framerate.
- Add support for non integer framerates (e.g. 23.976Hz, 59.94Hz etc)
- Support for nVidia Dynamic Twinview using their NV-CTRL X11 extension
- By default will try progressive refresh rates first. Interlaced video modes are disabled when using Dynamic Twinview for the time being. If you want to use an interlaced display with the automatic selection of the best refresh rate: you must disable Dynamic Twinview in your xorg.conf configuration.

To use, in Setup -> Appearance -> Video Mode Settings. Check "Separate video modes for GUI and TV playback" ; set the resolution of your screen there (best is to use the same GUI size and Video Ouput size) and set the value "Rate" to "Any". Mythtv will automatically try to find the best refresh rate supported by your TV when playing a video.

Note: Those changes uncovered a bug that existed previously when the change of refresh rate occur, the screen is not properly refreshed.


git-svn-id: http://svn.mythtv.org/svn/trunk@21489 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmyth/mythcontext.cpp
@@ -1653,7 +1653,7 @@ void MythContext::RefreshBackendConfig(void)
 void MythContext::GetResolutionSetting(const QString &type,
                                        int &width, int &height,
                                        double &forced_aspect,
-                                       short &refresh_rate,
+                                       double &refresh_rate,
                                        int index)
 {
     d->m_database->GetResolutionSetting(type, width, height, forced_aspect,
